// src/lib/types/super-admin/adminType.ts
interface AdminFormDataValueProps {
    name: string;
    username: string;
    password: string;
    email: string;
    departments: string[];
    address: string;
    origin: string
    organization: string
    gender: string;
    note: string;
    phone_number: string
    profile_url: string
    profile_key: string
}

export type { AdminFormDataValueProps }